requirements.txt
PySimpleGUI == 4.32.1
PyYAML == 5.3.1
opencv_python == 3.4.2.17
